use before-string instead of display for header
Using 'display was making the level 1 header bullets invisible.  Instead of
using display, we can use before-string, which will not affect the text under
the overlay.

The before string will not be formatted by org-tree-slide-header-overlay-face so
this will need to be replaced.  Customization of header text will be replaced in
a future commit to re-enable org-tree-slide-header-overlay-face