Actions track their own headings
Numerous element mapping methods were removed from the slide class.  Now actions
do the work mostly on their own.  They probably don't even need to talk to the
parent anymore.

Of particular improvement was making section-next and section-previous use this
local heading and the action's own marker as the position.  This means that the
point is not changed, and therefore we don't need save excursion and the child
can advance its marker when calling section-next on its own

All of the actions were greatly simplified in their implementation by this
improvement.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>