Narrowing is a slide action.  A tale of improvement.
Before this commit, there was an understanding in the first child action that if
the buffer had been narrowed to zero, the child was to handle all display.

This meant that if the restriction was not zero-length, the child would know it
was being shown together on one big slide, and it could include itself rather
than replace the buffer restriction.

Along came the inline slide, and it was found this degree of cooperation
afforded very little nuance, too little.

The ability to override the child actions and to supply extra arguments to
actions was introduced.  This allowed the inline child action to configure its
child slides more precisely.

The narrowing action was made more configurable, supporting most use cases for
adjusting the restriction.

Finally, the inline slide would supply this properly configured narrowing slide
action to its children.  In the future, they will supply it to their children.

And the cooperation over the buffer restriction was ended.  Hooray.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>