!dirty, pre-refactor commit to comment on architecture
These changes are exposing a change in value preference around fundamental
trade-offs for parent-child slide independence versus composition and especially
the effects on the user configuration API and the technical demands of even
basic composition.

Originally the design was intended to essentially "call" child slides.  Because
these calls didn't depend on the parent, the child action was to run after the
slide and section actions.  The child action would just return the slide to the
deck because theoretically it doesn't care after that.

However, it is increasingly clear that children are more sub-routine like and
must run within the context of the parent to support features like the
round-robin child or babel scripting in a child that wants to be run in the
context of the parent.  This is in addition to a repeatedly encountered dilemma
of whether to run an implied backward step after `ms-end', entirely a symptom of
narrowing and the parent-child relationship.

Two design goals favored child-after-slide design:
- `ms-goto' for point-based indexing
- `ms-end' to achieve backward movement into a previous sibling tree without
   playing forward all of that tree's contents

At present there are three things being juggled that come up against the choice
to prefer child-after-slide:

- The idea of a non-display babel only child heading seems great for scripting on
  top of some content in the parent.  However, this question of whether the child
  is run within the parent or after the parent came up again.

- This argument I was creating, :compose, in order to hack around this
  limitation finally became very fiddly.  The default child action is so brainless
  that it can't properly react to :compose anyway.

- The idea of returning `step' to tell the deck that an init or end method call
  should count as a step is a symptom of how a parent needs to narrow to its
  contents after a child "returns", but the narrowing is done in the
  `step-backward' method, so we need to call `step-backward' after switching
  back to a parent.  Not every action wants this implicit extra backward.  If
  narrowing was handled in the `ms-init' and `ms-end' and restored when calling
  `ms-final' of a child, the default narrowing action wouldn't need the extra step.

Worst of all, the effect of all these little fights is that running a child a
certain way requires configuration on both the parent and the child.  For the
user configuration, this is horrible.  It's completely non-intuitive and half of
the options and variables will be invalid except certain special combinations
that the user has to read to discover.

So the next set of commits will remove these sins and switch to
child-within-slide ordering of actions.  The child action always runs within the
lifecycle of the slide action, and the parent narrowing, if used, will cooperate
by consuming the forward / backward step after a child fails to make progress.
Because slides might compose, narrowing for inline children, which may have
several levels, will use overlays rather than narrowing, enabling deep
composition of revealing children.

Rather than relying on implied independence of children and parents, efficiently
implemented `ms-end' and `ms-goto' methods will save us from the worst possible
cases of replay when implementing backwards into sibling trees and goto from the
contents.  Then everything will be mostly pretty nice.

The downside is that a lot of know-how in the deck will find its way into
children.  This know-how can likely be generalized to reduce the pain, making
the deck inherit the know-how rather than the sole possessor of it.

Alright.  This plan makes sense.  Let's hope the next few commits reveals its
precision and accuracy.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>