Remove vestigal push-* code
The idea was that slides might push a restriction and pop it off the stack.
Because of how child action composition has to work, this was deemed not that
useful.

The deck no longer knows the depth of children, so it also cannot do this
properly.  Until child actions all play nice, this cannot be re-implemented.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>