Display strategy
First choice is usually going to be same buffer.  Second choice is in a previous
window that displayed the buffer.

For the deck, the previous window usually doesn't exist?  Usually when starting,
the intent is to re-use the window showing the org buffer.

Mainly want to allow the user to override via `display-buffer-alist'

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>