interesting bug in custom action
Due to editing this by hand, I accidentally did not include the if expression in
the when-let expression.  As a result, the if could return t forever, meaning it
was impossible to back up before this custom action.

Return values are sensitive.  No way to protect the user from it except to
hammer that fact home.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>