dslide-presentation -> dslide-deck
One less idea.  The deck is what the user commands talk to.  I could name it the
presentation or go with the shorter word, a deck.

PT etc is more common I think in non-English languages.  While this means more
people understand "presentation" it also mean's it's not a loved word due to its
complexity.

Dslide however, that's just fantastic.  Thanks for reading.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>