Bumping to 0.3.2 - expect about 95% name stability from here
Unless any of these names are wildly unpopular, the API is expected to be about
95% name-stable from here.  This is still not 1.0 software.While the potential
cost of not deprecating is going down, symbols are not yet being maintained.

I have done a lot of renaming, some of it demanded by conventions, some of it to
reduce the number of concepts or match typical Emacs patterns.

This was extremely important to get done ahead of submitting for MELPA.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>