introducing reverse-in-place
fairly important commit in terms of reaching maturity.  What has become clear is
that some actions want to reverse (calling forward and then backward or the
other way around) in place, repeating the same element.  Other actions do not.
Now they both have support.

The implementation difference is that when reversing in place, going forward
will set the marker to the end of the object.  Going backward still sets to the
beginning.  However the forward step matches even at the current marker
position, so after backward, the marker is at the beginning, and forward will
just repeat the last element that was processed going backward.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>