slide composition order change (soon to refactor)
There was a hack in place to allow the narrowing action to work.  This was
revealed to be buggy by embedding some section actions in a child.

Running the ordering in the originally intended way, with the slide always
encompassing the section and child actions, fixed an issue where nested slides
with items could behave bizarrely when going backward.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>