set overlay initial height
the animation was relying on the first frame of animation to first set the
height.  this ensures the overlay has the initial line height even if the timer
doesn't run before redisplay for some reason

Chasing flicker, also why I set a priority.. There were other sources of
flicker, and this can't all be confirmed useful yet.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>