version 0.4.0, some breaking changes
Child actions that were fiddled with for 0.3.x might break with 0.4.0, not that
any of this should have been stable.  It was very fiddly to write child actions
in 0.4.0

In 0.5.0, there is a strong chance that I combine child / slide action into just
multiple slide classes

Repeated development is revealing inherent coupling between the child and slide
action.  Their lifecycles are ambiguously encompassing and only the slide object
will be able to figure this out.

On the plus-side, it makes configuring the slides much, much easier.  Some of
the hydration code will die

The downside is that slide and child action won't be independent able to
articulate, but like I said, for the most part, they are not independent at all,
and what you want to do with the children is affected by what you do with the
slide, especially the restriction handling

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>