Remove child actions (pretty major refactor)
Basically all child actions needed to also control the restriction.  It became
apparent that handling the slide and the children is probably coupled in most
use cases.  The user would have to set multiple properties to achieve one
effect.  Furthermore, it was extremely tricky and remote to understand how the
slide and child actions were interacting.

This change alleviates this disparity by simply combining slide and child
actions into one action.

This is likely much easier even for very complex cases, such as bouncing between
the parent and a child or children in order to repeatedly make the connection
between several parts and a larger whole.

The changes to the demo.org demonstrate the simplicity that has been gained.
Architecturally, this approach will make it easier to implement nested actions.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>