! hide markup action
I didn't pretend the precedence or overrides are intact.  They need to be
overhauled.  To prevent the user from accidentally clobbering the default
actions, I made this always append for now.  Too many actions is usually less
bad than too few because actions are no-op if there's nothing to work on.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>