Off-by-one bug in markup hiding code
This was causing images with associated keywords to be hidden unless they were
preceded by a space, as seen in the demo.

It is unknown what forces of evil compelled me to include the extraneous space
before images in the demo.  Likely I encountered the issue while recording and
made it work, intending to double back and fix it before not understanding that
there was an issue when looking at my todo list.  Evil thrives in entropy and
can succeed at its nefarious goals even when it is miniscule and even while good
appears overwhelming.