Introducing Filtering
Using COMMENT or the :noslide: or :noexport: tags are supported.

Filtering has to be done on each child action and was implemented by updating
next and previous to skip filtered children.  The skipped children are hidden
after narrowing and can basically be ignored.

The contents must hide filtered headings.  Movement in the contents must skip
them.

Starting on a filtered heading will skip to the next heading or to a previous if
there is no next.