New Image Features!
- reveal images one by one
- images that only show in standalone

A remaining issue is that the behavior is not consistent forwards and backwards.
After showing the last image, we might want to show all of the images in the
sequence, and I'm not sure how that's possible yet.  WIP.

During this work, it was found that overlays that don't underlap the image
beginning by 1 don't hide the image.  Not sure what's going on there.  Was
unable to use dslide-hide-element, which works everywhere else.