! Prevent overlay leakage
At some point, I began recommending to users to just use `dslide-overlays' as a
dumping ground for overlays they wanted to be cleaned up when the slide was
"done".  However, these were not being cleaned up, resulting in duplicate
overlays.

`dslide-overlays' works really well for overlays that don't change throughout
the lifecycle of a slide.  The "correct" point to delete these is when the
restriction changes, usually meaning we're on a new slide.

If the new restriction is not included in the existing restriction, we cleanup
overlays.  However, this doesn't properly handle the case that the new
restriction contains the old restriction, meaning some of the existing overlays
may still be live.