Image feature cleanups.  Good to go 👌
So, this is the dirty make-it-all-work pass.

The code precisely embodies the challenges.

The image action uses both reverse-in-place and normal progress tracking
depending on how it's configured.  In order to both reveal images in the slide
and display them standalone, it needs to further use a step callback to hijack
the flow control.  This might not be strictly necessary, but it's what was done
for now.