Bug fix for reverse-in-place (!)
The manual section on progress tracking is must-read for working on this stuff.
The scheme for reverse-in-place is supposed to use the :end property when going
backwards.  Instead, it was only comparing the :begin property, resulting in
bugs that would affect actions that were very particular.

The first pass of the new image action features, with its callbacks and punching
the progress tracking marker, excited this failure mode.