Overhaul of start functions, introducing dslide-deck-develop
This is a big sloppy commit, I must regret.  It was supposed to be a small
change.  Unfortunately I had grown distant from this particular code because I
mainly just use `dslide-deck-start' and that was it.  Many of the functions have
unclear purpose and it's evident that the earlier changes to the lifecycle and
buffer management took their toll.  Rearrangement, renaming, and slamming the
new feature into place was just easier in the end.

Numerous issues were identified.  There will almost assuredly be some small
bugfixes here and there.  Sad.