fix kmacro direction & progress (was super broken)
Earlier today, I spotted a bug in copy-pasting backward from forward.
Unfortunately I "fixed" the forward case.

Anyway, I decided to properly implement progress.  The predicate was just
finding the keyword and not properly skipping blocks of the wrong direction, a
super broken behavior.

Now it is properly done.  Just about ready to stamp this as a release.