Handle buffer / window selection in image action, not command
The general direction is towards trusting the user to do the right thing with
regard to window selection and buffer movement.  Hand holding gets in the way of
advanced maneuvers.

The hand holding that existed was actually to support the image action, which
now has pretty okay logic to do the right thing