several tangled babel fixups
Most importantly, widening must be done before navigating to the block starting
position.  Otherwise, a final block in a heading with a child would fail to
evaluate properly.  This was introduced by me during this development cycle and
will not be mentioned in the news.

deprecated blocks were not able to be evaluated, again introduced during
development.

The user flow when ignoring a block failure was annoying because of the warning
display.  There is no longer any behavior of mapping over windows to move points
around.  This is handeled elsewhere and so added nothing but was risky in other
ways.