Use read-kbd-macro to make playback consistent
Had some issues, especially with characters like TAB not being interpreted
correctly.  This lead to a lengthy confirmation of a lot of behavior, ulimately
switching to `read-kbd-macro' and getting rid of the motivating issue without
much of a fight.

As a result, no translation is used and we are back at a more straightforward world.