Only abort playback on C-g.  WIP
During recording, I noticed certain last-input-events such as:

(file-notify ((30 . 0) (modify) "2024-12-29 21-00-31.mkv" 0)
file-notify--callback-inotify

These are not so helpful, but since I don't immediately know the range of other
such inputs, it was better to relax the checks.

C-g is supported.  Most users should know it and it will stop errant macros from
proceeding to destroy the system.

The comprehensive solution is to prevent fat-finger inputs in the first place.
It would be better to also support having the macro just dump remaining events
into unread-command-events whenever direction keys are called during playback.

I might still want to bind in the special map to block unwanted fat fingers.  It
is the only way to ensure that they don't corrupt the recorded sequence.